Added functionality for Incident
Added a check to see if the summary details contains a subject.  Subjects are created when an incident is created from an email.  If an incident is created from the api, the incident details will include a description

Method and support to incident class to retrieve the trigger log message.  The trigger log message carries extra information that the user sends in when an incident is created.  This informatino can be invaluable and now there is a simple method to retrieve it.