clean up oncalls and add tests
Added a unique id to the oncall container which ensures `list` works
with the Collection implemented pagination logic (requires an id).

Removed the overriden `show()` method which is inconsistent with how
show works on other objects (the `oncalls` endpoint does not support
getting a single oncall, as there is no PagerDuty id for them).